Button can cast ThunderStrike but TS hasnt work yet

+++ Assets/_Scripts/Enemy.cs
@@ -57,7 +57,7 @@ public class Enemy : MonoBehaviour {
         }
     }
 
-    public void TakeDamage(int damage) {
+    public void TakeDamage(float damage) {
         // Reduce health
         currentHealth -= damage;
